+++ ThrEd4/globals.h
@@ -62,7 +62,7 @@
 namespace fs = std::filesystem;
 
 extern uint32_t   ActiveColor;
-extern uint32_t   ActiveLayer;
+extern uint8_t   ActiveLayer;
 extern fRECTANGLE AllItemsRect;
 
 extern std::vector<fPOINT>* AngledFormVertices;
